+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 * Add `shorten` recipe for contemporary-math
+* Bake `key-terms` in `organic-chemistry`
+* Add class to `target-label`
+* Add `data-toc-target-page-type` attribute to TOC lis
+
 ## [v1.15.0] - 2022-08-25
 
 * Remove section subtitle from `Summary of Reaction` in `organic-chemistry`